updated folders and cleaned up my directory. removed files that were no longer needed or added to "old" folder for reference. ensured all steps in the code work - need to double-check syllabus upload.
