--- process_ChatGPT.py
@@ -1,20 +1,0 @@
-def process_syllabus(filepath):
-    with open(filepath, 'r') as file:
-        syllabus_text = file.read()
-
-    # Call ChatGPT API to process the syllabus
-    response = call_chatgpt_api(syllabus_text)
-    return response
-
-def call_chatgpt_api(text):
-    # Use the OpenAI to get a response
-    import openai
-    openai.api_key = 'your_openai_api_key'
-
-    response = openai.Completion.create(
-        model="text-davinci-003",
-        prompt=text,
-        max_tokens=1500
-    )
-
-    return response['choices'][0]['text']
