--- templates/google_ebook_reader_template.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html "-//W3C//DTD XHTML 1.0 Strict//EN"
-  "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-  <head>
-    <meta http-equiv="content-type" content="text/html; charset=utf-8"/>
-    <title>Google Books Embedded Viewer</title>
-    <link rel="stylesheet" href="/static/css/styles.css"> 
-    <script type="text/javascript" src="https://www.google.com/books/jsapi.js"></script>
-    <script type="text/javascript">
-      google.books.load();
-
-      function initialize() {
-        var viewer = new google.books.DefaultViewer(document.getElementById('viewerCanvas'));
-        viewer.load('ISBN:{{isbn}}');
-      }
-
-      google.books.setOnLoadCallback(initialize);
-    </script>
-  </head>
-  <body>
-    <div id="viewerCanvas" style="width: 1000px; height: 1000px"></div>
-  </body>
-</html>
