+++ templates/results.html
@@ -15,6 +15,9 @@
                     <strong>{{ book.title }}</strong> by {{ book.author }}<br>
                     {{ book.description }}<br>
                     <a href="{{ url_for('open_ebook', isbn=book.isbn, subject=book.subject, author=book.author) }}">Open Ebook</a>
+                    <a href="https://play.google.com/store/search?q={{ book.isbn }}&c=books" target="_blank">View on Google Play</a>
+                    <a href="https://www.amazon.com/s?k={{ book.isbn }}" target="_blank">View on Amazon</a>
+
                 </li>
             {% endfor %}
             </ul>
