--- templates_CH/google_ebook_reader_template.html.bak
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Ebook Reader</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-</head>
-<body>
-    <div class="container">
-        <h1>Reading Ebook with ISBN: {{ isbn }}</h1>
-        <!-- Your ebook reading logic goes here -->
-    </div>
-    <script src="{{ url_for('static', filename='js/scripts.js') }}"></script>
-</body>
-</html>
