--- templates_CH/results.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Search Results</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-</head>
-<body>
-    <div class="container">
-        <h1>Search Results</h1>
-        {% if books %}
-            <ul>
-            {% for book in books %}
-                <li>
-                    <strong>{{ book.title }}</strong> by {{ book.author }}<br>
-                    {{ book.description }}<br>
-                    <a href="{{ url_for('open_ebook', isbn=book.isbn, subject=book.subject, author=book.author) }}">Open Ebook</a>
-                    <a href="https://play.google.com/store/search?q={{ book.isbn }}&c=books" target="_blank">View on Google Play</a>
-                    <a href="https://www.amazon.com/s?k={{ book.isbn }}" target="_blank">View on Amazon</a>
-
-                </li>
-            {% endfor %}
-            </ul>
-        {% else %}
-            <p>No match found using search parameters</p>
-        {% endif %}
-        <a href="/">Back to search</a>
-    </div>
-    <script src="{{ url_for('static', filename='js/scripts.js') }}"></script>
-</body>
-</html>
