--- templates_CH/syllabus_results.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Syllabus Results</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-    <style>
-        .selected { background-color: lightgrey; } /* Highlight color for selected item */
-    </style>
-</head>
-<body>
-    <div class="container">
-        <h1>Syllabus Results</h1>
-        <p>Click on a textbook to select it</p>
-        <form id="searchForm" action="/search" method="post">
-            <input type="hidden" name="search_term" id="search_term" value="">
-            <input type="hidden" name="search_type" id="search_type" value="derived">
-            <input type="hidden" name="authors" id="authors" value="">
-            <input type="hidden" name="isbn" id="isbn" value="">
-            <ul>
-            {% for textbook in textbooks %}
-                <li onclick="selectTextbook('{{ textbook.title }}', '{{ textbook.authors }}', '{{ textbook.isbn }}', this)">
-                    {{ textbook.title }} by {{ textbook.authors }},  
-                    Publ: {{ textbook.publisher }},  
-                    Ed: {{ textbook.edition }}<br> 
-                    ISBN: {{ textbook.isbn }},  
-                    Requr.: {{ textbook.required }}, 
-                    Subj: {{ textbook.subject }},
-                </li>
-            {% endfor %}
-            </ul>
-            <button type="submit">Search for selected Textbook</button>
-        </form>
-        <a href="/">Back to search</a>
-    </div>
-    <script>
-        function selectTextbook(title, authors, isbn, element) {
-            // Remove selection from any previously selected item
-            document.querySelectorAll('li').forEach(li => li.classList.remove('selected'));
-            // Mark the current item as selected
-            element.classList.add('selected');
-    
-            // Update the hidden input values with the selected textbook's information
-            document.getElementById('search_term').value = title;
-            document.getElementById('authors').value = authors;
-            document.getElementById('isbn').value = isbn; // Corrected this line
-        }
-    </script>
