--- templates_CH/upload_syllabus.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Upload Syllabus</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/styles.css') }}">
-</head>
-<body>
-    <div class="container">
-        <h1>Upload Syllabus</h1>
-        <form action="{{ url_for('upload_syllabus') }}" method="post" enctype="multipart/form-data">
-            <input type="file" name="syllabus_pdf" accept="application/pdf">
-            <button type="submit">Upload</button>
-        </form>
-        <a href="/">Back to search</a>
-    </div>
-</body>
-</html>
